refactor(dialog): separate Dialog components into distinct files and implement context for better structure
