--- src/components/dialog/Dialog.tsx
@@ -1,168 +1,0 @@
-import React from "react";
-import { createPortal } from "react-dom";
-import { cn } from "@/lib/utils";
-import { useFocusTrap } from "@/hooks/useFocusTrap";
-import { Button } from "../common/Button";
-
-//WIP: separate concerns and components
-interface DialogRootProps {
-  open: boolean;
-  onOpenChange: (open: boolean) => void;
-  children: React.ReactNode;
-}
-
-const DialogContext = React.createContext<{
-  open: boolean;
-  onOpenChange: (open: boolean) => void;
-} | null>(null);
-
-export const Dialog: React.FC<DialogRootProps> = ({
-  children,
-  open,
-  onOpenChange,
-}) => {
-  return (
-    <DialogContext.Provider value={{ open, onOpenChange }}>
-      {children}
-    </DialogContext.Provider>
-  );
-};
-
-interface DialogTriggerProps {
-  asChild?: boolean;
-  children: React.ReactNode;
-}
-
-export const DialogTrigger: React.FC<DialogTriggerProps> = ({
-  asChild,
-  children,
-}) => {
-  const context = React.useContext(DialogContext);
-  if (!context) throw new Error("DialogTrigger must be used within Dialog");
-
-  const child = asChild ? (
-    React.Children.only(children)
-  ) : (
-    <Button>{children}</Button>
-  );
-
-  return React.cloneElement(child as React.ReactElement, {
-    onClick: () => context.onOpenChange(true),
-  });
-};
-
-interface DialogContentProps extends React.HTMLAttributes<HTMLDivElement> {
-  children: React.ReactNode;
-}
-
-export const DialogContent = React.forwardRef<
-  HTMLDivElement,
-  DialogContentProps
->(({ children, className, ...props }) => {
-  const context = React.useContext(DialogContext);
-  if (!context) throw new Error("DialogContent must be used within Dialog");
-  const { open, onOpenChange } = context;
-
-  const focusTrapRef = useFocusTrap(open, () => onOpenChange(false));
-  const previousFocus = React.useRef<HTMLElement | null>(null);
-
-  const handleBackdropClick = () => {
-    onOpenChange(false);
-  };
-
-  React.useEffect(() => {
-    if (open) {
-      previousFocus.current = document.activeElement as HTMLElement;
-      document.body.style.overflow = "hidden";
-    } else if (previousFocus.current) {
-      previousFocus.current.focus();
-      document.body.style.overflow = "";
-    }
-  }, [open]);
-
-  if (!open) return null;
-
-  return createPortal(
-    <div className="fixed inset-0 z-50" role="presentation">
-      <div
-        className={cn("fixed inset-0 backdrop-blur-sm", "animate-in fade-in-0")}
-        onClick={handleBackdropClick}
-        aria-hidden="true"
-      />
-      <div
-        ref={focusTrapRef}
-        role="dialog"
-        aria-modal="true"
-        aria-labelledby="dialog-title"
-        aria-describedby="dialog-description"
-        className={cn(
-          "fixed left-[50%] top-[50%] z-50 flex w-full max-w-xl translate-x-[-50%] translate-y-[-50%] gap-4",
-          "animate-in fade-in-0 zoom-in-95 slide-in-from-left-1/2 slide-in-from-top-[48%] rounded-lg bg-neutral-3 p-4 shadow-lg duration-200",
-          className,
-        )}
-        onClick={(e) => e.stopPropagation()}
-        {...props}
-      >
-        {children}
-      </div>
-    </div>,
-    document.body,
-  );
-});
-
-DialogContent.displayName = "DialogContent";
-
-export const DialogHeader = ({
-  className,
-  ...props
-}: React.HTMLAttributes<HTMLDivElement>) => (
-  <div
-    className={cn(
-      "flex flex-col space-y-1.5 text-center sm:text-left",
-      className,
-    )}
-    {...props}
-  />
-);
-
-export const DialogFooter = ({
-  className,
-  ...props
-}: React.HTMLAttributes<HTMLDivElement>) => (
-  <div
-    className={cn(
-      "flex flex-col-reverse sm:flex-row sm:justify-end sm:space-x-2",
-      className,
-    )}
-    {...props}
-  />
-);
-
-export const DialogTitle = React.forwardRef<
-  HTMLHeadingElement,
-  React.HTMLAttributes<HTMLHeadingElement>
->(({ className, ...props }, ref) => (
-  <h2
-    ref={ref}
-    className={cn(
-      "text-lg font-semibold leading-none tracking-tight",
-      className,
-    )}
-    {...props}
-  />
-));
-
-DialogTitle.displayName = "DialogTitle";
-
-export const DialogDescription = React.forwardRef<
-  HTMLParagraphElement,
-  React.HTMLAttributes<HTMLParagraphElement>
->(({ className, ...props }, ref) => (
-  <p
-    ref={ref}
-    className={cn("text-muted-foreground text-sm", className)}
-    {...props}
-  />
-));
-
-DialogDescription.displayName = "DialogDescription";
