+++ src/components/dialog/index.ts
@@ -1,9 +1,7 @@
-export {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-  DialogTrigger,
-} from "./Dialog";
+export { DialogRoot as Dialog } from "./primitives/DialogRoot";
+export { DialogTrigger } from "./primitives/DialogTrigger";
+export { DialogContent } from "./primitives/DialogContent";
+export { DialogHeader } from "./primitives/DialogHeader";
+export { DialogFooter } from "./primitives/DialogFooter";
+export { DialogTitle } from "./primitives/DialogTitle";
+export { DialogDescription } from "./primitives/DialogDescription";
