+++ src/components/dialog/primitives/DialogFooter.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+import { cn } from "@/lib/utils";
+
+export const DialogFooter = ({
+  className,
+  ...props
+}: React.HTMLAttributes<HTMLDivElement>) => (
+  <div
+    className={cn(
+      "flex flex-col-reverse sm:flex-row sm:justify-end sm:space-x-2",
+      className,
+    )}
+    {...props}
+  />
+);
