+++ src/components/dialog/primitives/DialogHeader.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+import { cn } from "@/lib/utils";
+
+export const DialogHeader = ({
+  className,
+  ...props
+}: React.HTMLAttributes<HTMLDivElement>) => (
+  <div
+    className={cn(
+      "flex flex-col space-y-1.5 text-center sm:text-left",
+      className,
+    )}
+    {...props}
+  />
+);
